ai-meeting-extractor/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.9.2
google-generativeai==0.7.2
pypdf==4.3.1
